moved like controller/service to the corresponding picture/comment controller/service following lukes suggestion
