+++ backend/src/main/java/com/soen341/instagram/controller/PictureController.java
@@ -55,4 +55,17 @@ public class PictureController {
             throw new AccountNotFoundException("The specified user could not be found");
         return pictureService.getAccountPictures(user);
     }
+    
+    @PostMapping(value = "/picture/like/{pictureId}")
+	public int likePicture(@PathVariable final long pictureId)
+	{
+		return pictureService.likePicture(pictureId);
+	}
+	
+	@PostMapping(value = "/picture/likeRemoval/{pictureId}")
+	public int unlikePicture(@PathVariable final long pictureId)
+	{
+		return pictureService.unlikePicture(pictureId);
+	}
+    
 }
