+++ backend/src/main/java/com/soen341/instagram/service/impl/CommentService.java
@@ -3,6 +3,7 @@ package com.soen341.instagram.service.impl;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -17,6 +18,7 @@ import com.soen341.instagram.dao.impl.PictureRepository;
 import com.soen341.instagram.exception.comment.CommentLengthTooLongException;
 import com.soen341.instagram.exception.comment.CommentNotFoundException;
 import com.soen341.instagram.exception.like.MultipleLikeException;
+import com.soen341.instagram.exception.like.NoLikeException;
 import com.soen341.instagram.exception.picture.InvalidIdException;
 import com.soen341.instagram.exception.picture.PictureNotFoundException;
 import com.soen341.instagram.model.Account;
@@ -126,4 +128,38 @@ public class CommentService
 		}
 	}
 
+	// like service
+
+	public int likeComment(final long commentId)
+	{
+		final Comment comment = findComment(commentId);
+		final Set<Account> likedBy = comment.getLikedBy();
+		final boolean addedSuccessfully = likedBy.add(getCurrentUser());
+		if (!addedSuccessfully)
+		{
+			throw new MultipleLikeException("A comment can only be liked once by the same user.");
+		}
+
+		commentRepository.save(comment);
+		System.out.println("\nYou have liked a comment. "+comment.getLikeCount());
+		return comment.getLikeCount();
+	}
+
+	public int unlikeComment(final long commentId)
+	{
+		final Comment comment = findComment(commentId);
+		final Set<Account> likedBy = comment.getLikedBy();
+		final boolean removedSuccessfully = likedBy.remove(getCurrentUser());
+
+		if (!removedSuccessfully)
+		{
+			throw new NoLikeException("The comment has not been liked by the user");
+		}
+
+		commentRepository.save(comment);
+
+		return comment.getLikeCount();
+	}
+	
+	
 }
